Refactor: Separate UI and logging logic into distinct files
- I separated JavaScript logic from the main `script.js` into `ui.js` and `logging.js`.
- `ui.js` now contains all DOM manipulation, UI update functions, event listeners, and UI-specific data (`userUploadedFolders`, `demoFilesData`).
- `logging.js` now contains core logging logic (`fileLog`), file hashing (`generateSHA256`), file processing (`scanFiles`), and log management functions (`loadLogsFromStorage`, `clearLogs`).
- I updated `index.html` to load `logging.js` before `ui.js`. The original `script.js` is kept but now only contains comments.
- I ensured proper global accessibility for shared variables (e.g., `window.fileLog`, `window.userUploadedFolders`) and functions to maintain interactivity between the two new files.
- This refactoring improves code organization, clarity, and maintainability of the GoldHash project's frontend codebase.
