+++ Projects/GoldHash/index.html
@@ -169,5 +169,7 @@
 </div>
 </div>
 
+<script src="logging.js"></script>
+<script src="ui.js"></script>
 <script src="script.js"></script>
 </body></html>
